requirements.txt
-i https://pypi.org/simple
annotated-types==0.6.0; python_version >= '3.8'
anyio==4.3.0; python_version >= '3.8'
certifi==2024.2.2; python_version >= '3.6'
click==8.1.7; python_version >= '3.7'
dnspython==2.6.1; python_version >= '3.8'
email-validator==2.1.1
exceptiongroup==1.2.1; python_version < '3.11'
fastapi[all]==0.110.2; python_version >= '3.8'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.5; python_version >= '3.8'
httptools==0.6.1
httpx==0.27.0
idna==3.7; python_version >= '3.5'
itsdangerous==2.2.0
jinja2==3.1.3
markupsafe==2.1.5; python_version >= '3.7'
orjson==3.10.1
pydantic==2.7.0; python_version >= '3.8'
pydantic-core==2.18.1; python_version >= '3.8'
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
python-dotenv==1.0.1; python_version >= '3.8'
python-multipart==0.0.9
pyyaml==6.0.1
sniffio==1.3.1; python_version >= '3.7'
starlette==0.37.2; python_version >= '3.8'
typing-extensions==4.11.0; python_version >= '3.8'
ujson==5.9.0
uvicorn[standard]==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
